Adds info regarding download corruption in nginx
The proxy_buffering directive defaults to on, so every response gets buffered.
Normally thats not an issue, but when a slow client tries to download a huge file the download gets corrupted.
To prevent this from happening, the buffering of responses from GoCD must be disabled.

(cherry picked from commit 2074e6355313355599acc690574751002a23c38f)